# apps/server/src/agents/agents.dto.ts
import { ApiProperty } from '@nestjs/swagger';

import { TIERS } from '@keys/domain';

/**
 * What the wire carries. For the OpenAPI document, not for validation — the
 * controllers validate and the domain refuses, and a DTO that also validated
 * would be a second place a rule lives.
 *
 * Read the request bodies here and notice there is no tier on any of them.
 * That is not an omission to be corrected later: there is nowhere in this
 * product for a claimed tier to go.
 */

export class SignUpBody {
  @ApiProperty({ example: 'Chinedu Okafor' })
  displayName!: string;

  @ApiProperty({ example: '+2348012345678' })
  phone!: string;
}

export class SignUpResponse {
  @ApiProperty()
  agentId!: string;

  @ApiProperty({ description: 'Shown once. Keys does not store it, only its digest.' })
  token!: string;
}

export class AgentProfileResponse {
  @ApiProperty()
  agentId!: string;

  @ApiProperty()
  displayName!: string;

  @ApiProperty({ enum: TIERS, description: 'Computed from evidence on every read. Not stored.' })
  tier!: string;

  @ApiProperty({ description: 'What was actually checked, in words a tenant could go and verify.' })
  meaning!: string;

  @ApiProperty({ description: 'Properties a landlord has confirmed they may let.' })
  confirmedProperties!: number;

  @ApiProperty({ format: 'date-time' })
  joinedAt!: string;

  @ApiProperty({ description: 'Upheld scam reports against this agent. Zero is not a clean bill of health.' })
  upheldReports!: number;
}

export class IdentityCheckBody {
  @ApiProperty({ description: 'The KYC vendor that ran the liveness and document check.' })
  vendor!: string;

  @ApiProperty({ description: "The vendor's own reference for the check, so it can be re-pulled." })
  reference!: string;

  @ApiProperty({ description: 'The agent the vendor checked.' })
  agentId!: string;
}

export class AuthorityRequestBody {
  @ApiProperty({ description: 'The property this authority would cover.' })
  propertyId!: string;

  @ApiProperty({ example: '+2348012345678', description: "The landlord's number. Not the agent's." })
  landlordPhone!: string;
}

export class ChallengeOpenedResponse {
  @ApiProperty({ description: 'Given to the landlord, not to the agent.' })
  challengeId!: string;

  @ApiProperty({ format: 'date-time' })
  expiresAt!: string;

  @ApiProperty({ description: 'Whether the text could actually be sent yet.' })
  delivered!: boolean;

  @ApiProperty()
  whatHappensNext!: string;
}

export class AnswerChallengeBody {
  @ApiProperty()
  challengeId!: string;

  @ApiProperty({ example: '049217', description: 'The six digits from the text.' })
  code!: string;
}

export class ChallengeAnsweredResponse {
  @ApiProperty()
  confirmed!: boolean;

  @ApiProperty({ description: 'Listings that went dark as a result. Empty on a grant.' })
  unpublishedListings!: string[];

  @ApiProperty()
  meaning!: string;
}

export class CreateListingBody {
  @ApiProperty()
  propertyId!: string;

  @ApiProperty({ example: '2 bedroom flat, Yaba' })
  title!: string;
}

export class ListingResponse {
  @ApiProperty()
  id!: string;

  @ApiProperty()
  propertyId!: string;

  @ApiProperty()
  title!: string;

  @ApiProperty({ type: String, nullable: true, format: 'date-time' })
  publishedAt!: string | null;
}